+++ src/components/play/player-contols.tsx
@@ -3,6 +3,7 @@ import { mdiChevronDown } from '@mdi/js';
 
 import { Icon } from 'solo-ui';
 
+import { THEME } from '../../const';
 import { Player, usePlayerName, usePlayerIcon } from '../../services/player';
 import { InstrumentCounts } from '../../services/instrument-utils';
 import { Instruments } from '../../services/instrument';
@@ -10,7 +11,6 @@ import { InstrumentControls } from './instrument-controls';
 import { Text } from '../shared/text';
 
 import './player-controls.css';
-import { THEME } from '../../const';
 
 interface Props {
     player: Player;